Добавлен класс GameScreen
