+++ js/data/get-score.test.js
@@ -1,5 +1,5 @@
 import assert from 'assert';
-import getScore from './get-score.js';
+import getScore from './get-score';
 
 describe(`Get player score`, () => {
   let answers;
