+++ js/dom-helpers/show-mistake.js
@@ -1,5 +1,5 @@
-import getMistakesHTML from '../views/get-mistakes-html.js';
-import headerPresenter from '../screens/header-presenter.js';
+import getMistakesHTML from '../views/get-mistakes-html';
+import headerPresenter from '../screens/header-presenter';
 
 export default () => {
   headerPresenter.view.element.querySelector(`.main-mistakes`).innerHTML = getMistakesHTML();
