+++ js/screen-data/get-data-lose.js
@@ -1,4 +1,4 @@
-import gameData from '../data/game-data.js';
+import gameData from '../data/game-data';
 
 export default (comparison) => {
   return {
