+++ js/screens/welcome-screen.js
@@ -1,15 +1,19 @@
-import showNextQuestion from '../control/show-next-question.js';
-import WelcomeView from '../views/welcome-view.js';
-import dataWelcome from '../screen-data/data-welcome.js';
+import WelcomeView from '../views/welcome-view';
+import dataWelcome from '../screen-data/data-welcome';
+import GameScreen from '../screens/game-screen';
 
 class WelcomeScreen {
   constructor() {
     this.view = new WelcomeView(dataWelcome);
   }
 
-  initialize() {
+  initialize(data) {
+    if (typeof data !== `undefined`) {
+      this.view.data = data;
+    }
+
     this.view.onPlayClick = () => {
-      showNextQuestion(true);
+      GameScreen.showQuestion(true);
     };
   }
 }
