+++ js/views/logo.js
@@ -1,3 +1,3 @@
-import gameData from '../data/game-data.js';
+import gameData from '../data/game-data';
 
 export default `<section class="logo" title="${gameData.GAME_TITLE}"><h1>${gameData.GAME_TITLE}</h1></section>`;
